Incremental rendering while server is running.
While the JBake server is running, periodically poll for updated file
system contents and incrementally update the output folder. Only
copy/render those contents that has or potentially has changed.

+++ src/main/java/org/jbake/app/Asset.java
@@ -58,18 +58,19 @@ public class Asset {
 			Arrays.sort(assets);
 			for (int i = 0; i < assets.length; i++) {
 				if (assets[i].isFile()) {
-					StringBuilder sb = new StringBuilder();
-					sb.append("Copying [" + assets[i].getPath() + "]...");
 					File sourceFile = assets[i];
 					File destFile = new File(sourceFile.getPath().replace(source.getPath()+File.separator+config.getString(ConfigUtil.Keys.ASSET_FOLDER), destination.getPath()));
+					
+					if (destFile.exists() && destFile.lastModified() >= sourceFile.lastModified()) {
+						// Prevent copying identical files over and over.
+						continue;
+					}
+					
 					try {
 						FileUtils.copyFile(sourceFile, destFile);
-						sb.append("done!");
-						LOGGER.info(sb.toString());
+						LOGGER.info("Copying [" + assets[i].getPath() + "]: done!");
 					} catch (IOException e) {
-						sb.append("failed!");
-						LOGGER.error(sb.toString(), e);
-						e.printStackTrace();
+						LOGGER.error("Copying [" + assets[i].getPath() + "]: failed!", e);
 						errors.add(e);
 					}
 				}
